requirements.txt
django
python-dotenv
psycopg2
djangorestframework
drf_yasg
pytest